[finish 162363557]modify model views tests

+++ app/api/v1/redflags/views.py
@@ -1,3 +1,4 @@
+"""Docstring for views"""
 from flask_restful import Resource
 from flask import jsonify, make_response, request
 from .models import RedFlagModule
@@ -11,22 +12,12 @@ class RedFlags(Resource):
         self.db = RedFlagModule()
 
     def post(self):
+        """method for posting a redflag"""
         
-        data = {
-            'createdOn' : datetime.datetime.utcnow(),
-            'createdBy' : request.json.get('createdBy', ""),
-            'type' : 'red-flags',
-            'location' : request.json.get('location', ""),
-            'status' : "Under Investigation",
-            'images' : request.json.get('images', ""),
-            'videos' : request.json.get('videos', ""),
-            'title' : request.json['title'],
-            'comment' : request.json.get('comment', "")
-        }
-        self.db.save(data)
+        self.db.save()
         
         success_message = {
-            'message' : 'Created red-flag record'
+            'message' : 'Sucessfully created a red-flag'
         }
 
         return make_response(jsonify({
@@ -35,20 +26,27 @@ class RedFlags(Resource):
         }), 201)
 
     def get(self):
-        self.db.get_all()      
+        """method for getting all redflags"""
+        self.db.get_all() 
         return make_response(jsonify({
             "status" : 200,
             "data" : self.db.get_all()
-        }), 200) 
+        }), 200)
 
 class RedFlag(Resource):
-    """docstring of RedFlag"""
+    """docstring for RedFlag"""
     def __init__(self):
         self.db = RedFlagModule()
         
 
     def get(self, redflag_id):
+        """method for getting redflag by id"""
         incident = self.db.find(redflag_id)
+        if incident == None:
+            return make_response(jsonify({
+                "status" : 200,
+                "error" : "Red-flag does not exist"
+            }), 200)
         return make_response(jsonify({
                     "status" : 200,
                     "data" : incident
@@ -56,17 +54,26 @@ class RedFlag(Resource):
     
 
     def delete(self, redflag_id):
+        """method for deleting a single redflag by id"""
         incident = self.db.find(redflag_id)
-        self.db.delete(incident)
-        success_message = {
-                'message' : 'Red-flag record has been sucessfully deleted'
+
+        if incident == None:
+            return make_response(jsonify({
+                "status" : 200,
+                "data" : "Red flag does not exist"
+            }), 200)
+        delete_status = self.db.delete(incident)
+        if delete_status == "deleted":
+            return make_response(jsonify({
+                "status" : 200,
+                "data" : {
+                    "id" : redflag_id,
+                    "message" : 'Red-flag record has been sucessfully deleted'
                  }
-        return make_response(jsonify({
-                "status" : 204,
-                "data" : success_message
-                }))  
+            }), 200)  
 
     def put(self, redflag_id):
+        """method for putting redflag"""
         incident = self.db.find(redflag_id)
         if incident:
                 incident['createdBy'] = request.json.get('createdBy', incident['createdBy'])
@@ -86,36 +93,61 @@ class RedFlag(Resource):
                 }), 201)
 
 class UpdateLocation(Resource):
+    """Docstring for update location of a redflag"""
     def __init__(self):
         self.db = RedFlagModule()
 
 
     def patch(self, redflag_id):
+        """method for patching location"""
         incident = self.db.find(redflag_id)
-        if incident:
-            incident['location'] = request.json.get('location', incident['location'])
-            success_message = {
-                        "message" : "Sucessfully updated red-flag location"
-                    }
 
+        if incident == None:
             return make_response(jsonify({
-                "status" : 201,
-                "data" : success_message
-            }), 201)  
+                "status" : 200,
+                "error" : "Red-flag does not exist"
+            }), 200)
+        edit_status = self.db.edit_redflag_location(incident)
+        if edit_status == "keyerror":
+            return make_response(jsonify({
+                "status" : 400,
+                "data" : "Keyerror: location for the redflag not updated"
+            }), 400)
+        elif edit_status == "location updated":
+            return make_response(jsonify({
+                "status" : 200,
+                "data" : {
+                    "id" : redflag_id,
+                    "message" : "Successfully updated redflag location"
+                }
+            }), 200)
 
 class UpdateComment(Resource):
+    """Docstring for update location of a redflag"""
     def __init__(self):
         self.db = RedFlagModule()
+
+
     def patch(self, redflag_id):
+        """method for patching location"""
         incident = self.db.find(redflag_id)
-        if incident:
-                incident['comment'] = request.json.get('comment', incident['comment'])
 
-                success_message = {
-                    "message" : "Successfully updated red-flag comment"
+        if incident == None:
+            return make_response(jsonify({
+                "status" : 200,
+                "error" : "Red-flag does not exist"
+            }), 200)
+        edit_status = self.db.edit_redflag_comment(incident)
+        if edit_status == "keyerror":
+            return make_response(jsonify({
+                "status" : 400,
+                "data" : "Keyerror: comment for the redflag not updated"
+            }), 400)
+        elif edit_status == "comment updated":
+            return make_response(jsonify({
+                "status" : 200,
+                "data" : {
+                    "id" : redflag_id,
+                    "message" : "Successfully updated redflag comment"
                 }
-
-                return make_response(jsonify({
-                    "status" : 201,
-                    "data" : success_message
-                }), 201)                                                             
+            }), 200)
